tests: semaphore: consolidate semaphore tests
No need to have two tests for the same thing, just put everything in one
test.
