--- tests/kernel/semaphore/sema_api/prj.conf
@@ -1,4 +1,0 @@
-CONFIG_ZTEST=y
-CONFIG_IRQ_OFFLOAD=y
-CONFIG_TEST_USERSPACE=y
-CONFIG_MP_NUM_CPUS=1
