--- tests/kernel/semaphore/sema_api/testcase.yaml
@@ -1,3 +1,0 @@
-tests:
-  kernel.semaphore.api:
-    tags: kernel userspace
